refactored editRoutePathLayer into 3 files
